requirements.txt
Flask==3.1.0
gunicorn==20.1.0
requests~=2.32.3
protobuf~=5.29.3
beautifulsoup4~=4.12.3
google-generativeai==0.8.4
Flask-Cors~=5.0.0
python-dotenv~=1.0.1
aiohttp==3.9.1
asyncio==3.4.3
